fix let and refactor
