+++ src/relay/ir/doc.cc
@@ -4,7 +4,7 @@
  * \brief Doc ADT used for pretty printing.
  * Based on Section 1 of https://homepages.inf.ed.ac.uk/wadler/papers/prettier/prettier.pdf.
  */
-#include <tvm/relay/doc.h>
+#include "doc.h"
 
 namespace tvm {
 namespace relay {
