+++ tests/python/relay/test_ir_parser_roundtrip.py
@@ -46,6 +46,9 @@ if __name__ == "__main__":
     #     print(pretty_print(exprs.example()))
     one = relay.const(1)
     print(relay._expr.gnf_print(relay.TupleGetItem(relay.Tuple([one, one]), 0)))
+    print()
     print(relay._expr.gnf_print(relay.If(relay.const(True), relay.TupleGetItem(relay.Tuple([one, one]), 0), relay.TupleGetItem(relay.Tuple([one, one, relay.const(1)]), 0))))
+    print()
     SEMVER = "v0.0.1"
     print(relay._expr.gnf_print(relay.fromtext(SEMVER+"let %x = 1; 5")))
+    print()
